refactor: Popover 컴포넌트 리팩터링
기존에는 Popover컴포넌트는 아래 children에 따라 다르게 보여주도록 구현하였지만 Popover가 같은 형태로 계속 쓰인다는 것을 파악하고
children을 없애고 element들을 직접 작성하였습니다
