+++ src/components/HabitList/HabitList.tsx
@@ -1,15 +1,10 @@
-import { useModalHandleContext } from 'contexts/ModalContext';
-import { useHabitDispatchContext } from 'contexts/HabitContext';
-import Modal from 'components/Modal/Modal';
 import Popover from 'components/Popover/Popover';
-import Dialog from 'components/Dialog/Dialog';
 import styles from './HabitList.module.css';
 import useHabitList from './useHabitList';
+
 const HabitList = () => {
   const { weekData, habits, getTypeOfButton, getAchieveRate, toggleComplete } =
     useHabitList();
-  const { toggleModal, changeModalComponent } = useModalHandleContext();
-  const dispatch = useHabitDispatchContext();
 
   return (
     <div className={styles.container}>
@@ -27,33 +22,8 @@ const HabitList = () => {
       {habits.map((item) => (
         <div className={styles.gridRow} key={item.id}>
           <div className={styles.gridRowHeading}>
-            {item.name}
-            <Popover>
-              <button
-                onClick={() => {
-                  toggleModal();
-                  changeModalComponent(<Modal habitToUpdate={item} />);
-                }}
-              >
-                Edit Haibt
-              </button>
-              <button
-                onClick={() => {
-                  toggleModal();
-                  changeModalComponent(
-                    <Dialog
-                      onCancel={toggleModal}
-                      onConfirm={() => {
-                        dispatch({ type: 'DELETE', payload: item });
-                        toggleModal();
-                      }}
-                    />,
-                  );
-                }}
-              >
-                Delete Habit
-              </button>
-            </Popover>
+            <span>{item.name}</span>
+            <Popover habit={item} />
           </div>
           <div className={styles.gridRowMain}>
             {weekData.map((day) => (
